chore: pr review fixes
